moved styles into own folder
