+++ frontend/src/components/Navbar.tsx
@@ -1,5 +1,5 @@
 // import { useState, useEffect } from "react";
-import "./Navbar.css";
+import "../styles/Navbar.css";
 import AppBar from "@mui/material/AppBar";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
