+++ frontend/src/styles/CalendarBase.scss
